block_duplicate_release_deployment.groovy
Artifactory Plugin for blocking the release redeployment


This page contains the logic which should be implemented using a Groovy script . This logic is for the blocking the redeployment of maven release artifacts in Artifactory. We are not checking for other type of artifacts (like bower , npm , nuget ) because people
have overwrite/delete access on the permission targets for them


Logic for the Plugin/Script
1. Script should check whether the user (from which deployment request is coming) is admin or not ?
if (security.currentUser().isAdmin()) return

a) If YES , Then do not proceed any further and let the deployment happen.
b) If NO , then proceed to step 2.


2. Script should check if the repo type is maven and it handles releases only.
def repokey = request.getRepoPath().getRepoKey()
if ( repositories.getProperty(repokey, Type) = maven && repokey.getType() = releases )

a) If YES , then proceed to step 3.
b) If NO , Then do not proceed any further and let the deployment happen

3 . Script should check whether the artifact already exist or NOT (gavc path of the maven release artifact (which is going to be deployed already exists) . it should check the artifact only in the corresponding read only local repo (inhouse-local-ro will be the read-only repo for
inhouse-local)
If YES - Block the deployment
If NO - go ahead with the deployment.